requirements.txt
matrix_client >= 0.4.0
Mastodon.py >= 1.5.0
